Fondo css oscurecido, padding arreglado
