+++ src/components/fondo/fondo.css
@@ -49,16 +49,6 @@
 
 }
 
-.fondobox.formulario{
-    padding: 0px;
-    width: 100%;
-    height: auto;
-    background: rgb(47, 47, 47);
-    background: linear-gradient(210deg, rgb(28 28 28) 0%, rgb(14 14 14) 50%, rgb(30 30 30) 100%);
-    background-size: cover;
-    background-size: 100% 100%;
-}
-
 .fondobox.footer {
 
     position: absolute;
